Move: pages => app migration
