+++ src/app/auth/signIn/page.tsx
@@ -1,5 +1,7 @@
+'use client';
+
 import React from 'react';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 import { css } from '@emotion/react';
 
 // react-query
@@ -18,7 +20,7 @@ import {
 } from '@/components/Auth/@subComponents';
 
 // emotion
-import { signInMainStyle } from './SignIn.style';
+import { signInMainStyle } from './style';
 
 // constant
 import { SIGN_IN_INPUT_DATA, OAUTH_BUTTONS } from './constants';
