+++ src/app/auth/signUp/style.tsx
