+++ src/app/main/page.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { css } from '@emotion/react';
 
 // components
