+++ src/components/Auth/@subComponents/AuthContainer/AuthHeader/AuthHeader.tsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 
 import { Typography } from 'nimble-ds';
