+++ src/components/Landing/LandingContainer/LandingContainer.tsx
@@ -1,5 +1,5 @@
 import Image from 'next/image';
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 import { css } from '@emotion/react';
 
 // hooks
