+++ src/components/Main/Navigation/Navigation.tsx
@@ -1,4 +1,4 @@
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 import { css } from '@emotion/react';
 import Cookies from 'js-cookie';
