+++ src/components/Ui/Header/Header.tsx
@@ -1,4 +1,4 @@
-import { useRouter } from 'next/router';
+import { useRouter } from 'next/navigation';
 
 // components
 import { Button, Typography, FlexContainer } from 'nimble-ds';
