--- src/pages/404.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Custom404 = () => {
-    return <h1>404 - Page Not Found</h1>;
-};
-
-export default Custom404;
