--- src/pages/500.tsx
@@ -1,7 +1,0 @@
-import React from 'react';
-
-const Custom500 = () => {
-    return <h1>500 - Server-side error occurred</h1>;
-};
-
-export default Custom500;
