--- src/pages/_app.tsx
@@ -1,27 +1,0 @@
-import React from 'react';
-import '@/styles/globals.css';
-import type { AppProps } from 'next/app';
-
-import { QueryClient, QueryClientProvider } from 'react-query';
-import { RecoilRoot } from 'recoil';
-
-import '../assets/font/font.css';
-
-const queryClient = new QueryClient({
-    defaultOptions: {
-        queries: {
-            refetchOnWindowFocus: false,
-            retry: false
-        }
-    }
-});
-
-export default function App({ Component, pageProps }: AppProps) {
-    return (
-        <RecoilRoot>
-            <QueryClientProvider client={queryClient}>
-                <Component {...pageProps} />
-            </QueryClientProvider>
-        </RecoilRoot>
-    );
-}
