--- src/pages/auth/signIn/index.ts
@@ -1,3 +1,0 @@
-import SignIn from './SignIn';
-
-export default SignIn;
