--- src/pages/auth/signUp/index.ts
@@ -1,3 +1,0 @@
-import SignUp from './SignUp';
-
-export default SignUp;
